+++ examples/nested_dependency/build.py
@@ -10,10 +10,8 @@ cppstdlib = sbuildr.Library("stdc++")
 project = sbuildr.Project()
 
 # Assumes that single_dependency is located in nested_dependency/../single_dependency
-# CopyFetcher does not support the version parameter, so we provide no version here.
-# For other fetchers, the correct version should be provided.
 single_dependency_path = os.path.abspath(os.path.join(pathlib.Path.home(), "Python", "SBuildr", "examples", "single_dependency"))
-single_dependency = sbuildr.dependencies.Dependency(fetchers.CopyFetcher(single_dependency_path), builders.SBuildrBuilder(), version="")
+single_dependency = sbuildr.dependencies.Dependency(fetchers.CopyFetcher(single_dependency_path), builders.SBuildrBuilder())
 
 # Add targets for this project
 project.interfaces(["doubleDep.hpp"])